Update zh.yml
Replace 比特币现金 with 比特现金 after suggested it was more natural to say and of current use in Chinese forums. It was also noted that this short version has the potential risk of being wrongly translated back into a corruption like "bitcash" (direct translation) or "bcash" (direct translation plus a touch of malice).